Implemented base services. Renamed _Internal namespace to Internal in order to solve issue reported by SCA Renamed folder and namespace Interface to Interfaces in order to solve issue reported by SCA
