+++ SignNow.Net/Interfaces/IUserService.cs
@@ -2,7 +2,7 @@ using SignNow.Net.Model;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SignNow.Net.Interface
+namespace SignNow.Net.Interfaces
 {
     public interface IUserService
     {
