+++ SignNow.Net/Service/WebClientBase.cs
@@ -1,13 +1,14 @@
 
-using SignNow.Net._Internal.Service;
-using SignNow.Net.Interface;
+using SignNow.Net.Internal.Service;
+using SignNow.Net.Interfaces;
 using System;
 
 namespace SignNow.Net.Service
 {
-    public abstract class WebClientBase
+    public class WebClientBase
     {
         protected ISignNowClient SignNowClient { get; private set; }
+        protected Uri ApiBaseUrl { get; set; }
         public WebClientBase(Uri apiBaseUrl) : this(apiBaseUrl, null)
         {
 
@@ -15,6 +16,7 @@ namespace SignNow.Net.Service
 
         protected WebClientBase(Uri apiBaseUrl, ISignNowClient signNowClient)
         {
+            ApiBaseUrl = apiBaseUrl;
             SignNowClient = signNowClient ?? new SignNowClient();
         }
     }
