+++ SignNow.Net/_Internal/Service/SignNowClient.cs
@@ -1,11 +1,11 @@
 
-using SignNow.Net.Interface;
+using SignNow.Net.Interfaces;
 using SignNow.Net.Model;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SignNow.Net._Internal.Service
+namespace SignNow.Net.Internal.Service
 {
     class SignNowClient : ISignNowClient
     {
